requirements.txt
Django==1.8.3

dj-database-url==0.3.0
psycopg2==2.6.1
project-runpy==0.3.1
django-extensions==1.5.5
django-taggit==0.16.2


# Deployment
whitenoise==2.0.2
waitress==0.8.9


# Dev requirements
factory-boy==2.5.2
mock==1.2.0
